+++ src/Readers/JetReader.cpp
@@ -59,15 +59,7 @@ JetReader::JetReader(TChainPointer input, JetAlgorithm::value algo) :
     btagSimpleSecondaryVertexReader(input, JetAlgorithm::prefixes.at(algo) + ".SimpleSecondaryVertexHighEffBTag"),
     btagTrackCountingHighPurityReader(input, JetAlgorithm::prefixes.at(algo) + ".TrackCountingHighPurBTag"),
     btagTrackCountingHighEfficiencyReader(input, JetAlgorithm::prefixes.at(algo) + ".TrackCountingHighEffBTag"),
-//    genEnergyReader(input, "GenJet.Energy"),
-//    genPxReader(input, "GenJet.Px"),
-//    genPyReader(input, "GenJet.Py"),
-//    genPzReader(input, "GenJet.Pz"),
-//    genMassReader(input, "GenJet.Mass"),
-//    genChargeReader(input, "GenJet.Charge"),
-//    genEmfReader(input, "GenJet.EMF"),
     jets(),
-//    genJets(),
     usedAlgorithm(algo) {
 
 }
@@ -81,19 +73,12 @@ const JetCollection& JetReader::getJets() {
 	return jets;
 }
 
-//const JetCollection& JetReader::getGenJets() {
-//	if (genJets.empty() == false)
-//		genJets.clear();
-//	readGenJets();
-//	return genJets;
-//}
-
 void JetReader::readJets() {
 	for (unsigned int jetIndex = 0; jetIndex < energyReader.size(); jetIndex++) {
-		float energy = energyReader.getVariableAt(jetIndex);
-		float px = pxReader.getVariableAt(jetIndex);
-		float py = pyReader.getVariableAt(jetIndex);
-		float pz = pzReader.getVariableAt(jetIndex);
+		double energy = energyReader.getVariableAt(jetIndex);
+		double px = pxReader.getVariableAt(jetIndex);
+		double py = pyReader.getVariableAt(jetIndex);
+		double pz = pzReader.getVariableAt(jetIndex);
 		JetPointer jet(new Jet(energy, px, py, pz));
 		jet->setUsedAlgorithm(usedAlgorithm);
 		jet->setMass(massReader.getVariableAt(jetIndex));
@@ -132,23 +117,6 @@ void JetReader::readJets() {
 	}
 }
 
-//void JetReader::readGenJets() {
-//	for (unsigned int jetIndex = 0; jetIndex < genEnergyReader.size(); jetIndex++) {
-//		float energy = genEnergyReader.getVariableAt(jetIndex);
-//		float px = genPxReader.getVariableAt(jetIndex);
-//		float py = genPyReader.getVariableAt(jetIndex);
-//		float pz = genPzReader.getVariableAt(jetIndex);
-//		JetPointer jet(new Jet(energy, px, py, pz));
-//
-//		jet->setUsedAlgorithm(usedAlgorithm);
-//		jet->setMass(genMassReader.getVariableAt(jetIndex));
-//		jet->setCharge(genChargeReader.getVariableAt(jetIndex));
-//		jet->setEMF(genEmfReader.getVariableAt(jetIndex));
-//
-//		genJets.push_back(jet);
-//	}
-//}
-
 void JetReader::initialise() {
 	energyReader.initialise();
 	pxReader.initialise();
@@ -173,13 +141,5 @@ void JetReader::initialise() {
 		CHFReader.initialise();
 		NCHReader.initialise();
 	}
-
-//	genEnergyReader.initialiseBlindly();
-//	genPxReader.initialiseBlindly();
-//	genPyReader.initialiseBlindly();
-//	genPzReader.initialiseBlindly();
-//	genMassReader.initialiseBlindly();
-//	genEmfReader.initialiseBlindly();
-//	genChargeReader.initialiseBlindly();
 }
 }
